jrcp Refactor
